requirements.txt
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
pkg-resources==0.0.0
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1